MINIDUMP: Fix compiler warnings

+++ kernel/printk/printk.c
@@ -1922,7 +1922,7 @@ EXPORT_SYMBOL(printk);
 #ifdef CONFIG_QCOM_MINIDUMP
 void get_log_buf_info(uint64_t *plog_buf, uint64_t *plog_buf_len)
 {
-       *plog_buf = log_buf;
+       *plog_buf = (uint64_t)log_buf;
        *plog_buf_len = (uint64_t)__pa(&log_buf_len);
 }
 #endif
